+++ site/doc.js
@@ -231,7 +231,7 @@ const page = t("html", {}, [
 				t("p", { class: "title", }, "Usage"),
 				t("p", { class: "desc", }, "quick start"),
 				code(`
-<script src="https://kaboomjs.com/lib/dev/kaboom.js"></script>
+<script src="https://kaboomjs.com/lib/0.1.0/kaboom.js"></script>
 <script type="module">
 
 // make kaboom functions global
@@ -253,6 +253,7 @@ scene("main", () => {
 
 // start the game
 start("main");
+
 </script>
 				`, "html"),
 				...api.map((sec) => {